+++ packages/atomic/src/components/search/result-template-components/atomic-result-number/atomic-result-number.new.stories.tsx
@@ -1,8 +1,13 @@
-import type {Meta, StoryObj as Story} from '@storybook/web-components';
+import type {Meta, StoryObj as Story} from '@storybook/web-components-vite';
+import {getStorybookHelpers} from '@wc-toolkit/storybook-helpers';
 import {parameters} from '@/storybook-utils/common/common-meta-parameters';
-import {renderComponent} from '@/storybook-utils/common/render-component';
 import {wrapInResult} from '@/storybook-utils/search/result-wrapper';
 import {wrapInSearchInterface} from '@/storybook-utils/search/search-interface-wrapper';
+
+const {events, args, argTypes, template} = getStorybookHelpers(
+  'atomic-result-number',
+  {excludeCategories: ['methods']}
+);
 
 const {decorator: resultDecorator, engineConfig} = wrapInResult({
   preprocessRequest: (r) => {
@@ -13,17 +18,25 @@ const {decorator: resultDecorator, engineConfig} = wrapInResult({
     return r;
   },
 });
-const {decorator: searchInterfaceDecorator, play} =
+const {decorator: searchInterfaceDecorator, afterEach} =
   wrapInSearchInterface(engineConfig);
 
 const meta: Meta = {
   component: 'atomic-result-number',
   title: 'Search/ResultList/ResultNumber',
   id: 'atomic-result-number',
-  render: renderComponent,
+  render: (args) => template(args),
   decorators: [resultDecorator, searchInterfaceDecorator],
-  parameters,
-  play,
+  parameters: {
+    ...parameters,
+    actions: {
+      handles: events,
+    },
+  },
+  args,
+  argTypes,
+
+  afterEach,
 };
 
 export default meta;
@@ -31,7 +44,7 @@ export default meta;
 export const Default: Story = {
   name: 'atomic-result-number',
   args: {
-    'slots-default': `
+    'default-slot': `
       <span style="font-weight: bold; margin-right: 0.25rem;">File size:</span>
       <atomic-format-unit unit="byte" unit-display="long"></atomic-format-unit>
     `,